feat: Config typed builder
