+++ Cargo.toml
@@ -23,3 +23,4 @@ embedded-hal = "0.2"
 embedded-hal-async = { version = "1.0.0-rc.1", optional = true }
 
 num-traits = { version = "0.2", default-features = false, features = ["libm"] }
+typed-builder = "0.18"
